Add files via upload
Just the basic functionality of the chat bot. Will update the code to try and make it integrate able with the main banking landing page and pop up in the bottom of the screen. At the moment I have created a html page to play around with the chat bot. 
Will also update the chat bot questions and answers to relate to financial advice and make it more fun. 
Also experimenting with pre set questions, will update that too.
I am still getting used to forks, branches and pull request so I apologise if this is not the right way to do it.